+++ DemoApp/StreamChat/Components/DemoChatMessageActionsVC.swift
@@ -12,12 +12,20 @@ final class DemoChatMessageActionsVC: ChatMessageActionsVC {
     override var messageActions: [ChatMessageActionItem] {
         var actions = super.messageActions
         if message?.isSentByCurrentUser == true {
-            actions.append(pinMessageActionItem())
+            
+            if message?.isBounced == false {
+                actions.append(pinMessageActionItem())
+            }
+            
             if AppConfig.shared.demoAppConfig.isHardDeleteEnabled {
                 actions.append(hardDeleteActionItem())
             }
         }
-        actions.append(translateActionItem())
+        
+        if message?.isBounced == false {
+            actions.append(translateActionItem())
+        }
+        
         return actions
     }